requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
slowapi==0.1.7
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==1.10.15
